+++ src/global/integrations/index.js
@@ -1,0 +1,8 @@
+import { createEngramProvider } from "./engram-provider.js";
+import { registerIntegrationProvider, getIntegrationProvider } from "./provider-registry.js";
+
+export function ensureIntegrationProvidersRegistered() {
+  if (!getIntegrationProvider("engram")) {
+    registerIntegrationProvider(createEngramProvider());
+  }
+}